requirements.txt
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
langchain-community==0.0.10
langchain-core>=0.1.7,<0.2
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.9.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
boto3>=1.34.0
botocore>=1.34.0